theme code separated into own file
